pyproject.toml
[tool.poetry]
name = "mlflow-openai"
version = "0.1.0"
description = ""
authors = ["Neelesh Batham <neelesh.batham007@gmail.com>"]
readme = "README.md"
packages = [{include = "mlflow_openai"}]

[tool.poetry.dependencies]
python = "^3.11"
mlflow = "^2.8.1"
gunicorn = "^21.2.0"
numpy = "^1.26.2"
openai = "^1.3.8"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
